add: Docs for the NIMFS of the scheme and related structs
